Put ERP data class into its own file, changed imports for files that use ERP data
