+++ examples/p300_ch_select_test.py
@@ -5,7 +5,7 @@ Test P300 offline using data from an existing stream
 
 import os
 
-from bci_essentials.bci_data import ERP_data
+from bci_essentials.erp_data import ERP_data
 from bci_essentials.classification.erp_rg_classifier import ERP_rg_classifier
 
 # Identify the file to simulate
