+++ tests/test_classifier_save.py
@@ -2,7 +2,8 @@ import unittest
 import os
 import numpy as np
 
-from bci_essentials.bci_data import EEG_data, ERP_data
+from bci_essentials.bci_data import EEG_data
+from bci_essentials.erp_data import ERP_data
 from bci_essentials.classification.mi_classifier import MI_classifier
 from bci_essentials.classification.erp_rg_classifier import ERP_rg_classifier
 from bci_essentials.session_saving import save_classifier, load_classifier
