+++ tests/test_smoke.py
@@ -1,7 +1,8 @@
 import unittest
 import os
 
-from bci_essentials.bci_data import EEG_data, ERP_data
+from bci_essentials.bci_data import EEG_data
+from bci_essentials.erp_data import ERP_data
 from bci_essentials.classification.mi_classifier import MI_classifier
 from bci_essentials.classification.erp_rg_classifier import ERP_rg_classifier
 from bci_essentials.classification.ssvep_riemannian_mdm_classifier import (
